% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/job.R
\name{AddArguments}
\alias{AddArguments}
\title{Add one or more arguments to the job}
\usage{
AddArguments(job, arguments)
}
\arguments{
\item{job}{Job object}

\item{arguments}{List of arguments defined as \code{list()}}
}
\value{
Job with appended list of arguments
}
\description{
Add one or more arguments to the job (this will add whitespace).
}
\seealso{
\code{\link{Job}}
}

