% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{AddDAG}
\alias{AddDAG}
\title{Add a sub-DAG (synonym for addJob)}
\usage{
AddDAG(adag, dag)
}
\arguments{
\item{adag}{ADAG object}

\item{dag}{Sub-DAG to be appended}
}
\value{
The ADAG object with the sub-DAG appended
}
\description{
Add a sub-DAG (synonym for addJob)
}
\seealso{
\code{\link{ADAG}}, \code{\link{AddJob}}, \code{\link{DAG}}
}

