% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{AddDAX}
\alias{AddDAX}
\title{Add a sub-DAX (synonym for addJob)}
\usage{
AddDAX(adag, dax)
}
\arguments{
\item{adag}{ADAG object}

\item{dax}{Sub-DAX to be appended}
}
\value{
The ADAG object with the sub-DAX appended
}
\description{
Add a sub-DAX (synonym for addJob)
}
\seealso{
\code{\link{ADAG}}, \code{\link{AddJob}}, \code{\link{DAX}}
}

