% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{AddDependency}
\alias{AddDependency}
\title{Add a dependency to the workflow}
\usage{
AddDependency(adag, dep)
}
\arguments{
\item{adag}{The ADAG object}

\item{dep}{The dependency object}
}
\value{
The ADAG object containing the dependency
}
\description{
Add a dependency to the workflow
}
\seealso{
\code{\link{ADAG}}, \code{\link{Dependency}},
 \code{\link{Depends}}, \code{\link{RemoveDependency}}
}

