% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{AddExecutable}
\alias{AddExecutable}
\title{Add an executable to the ADAG}
\usage{
AddExecutable(adag, executable)
}
\arguments{
\item{adag}{ADAG object}

\item{executable}{Executable object}
}
\value{
The ADAG object with the executable appended
}
\description{
Add an executable to the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{Executable}},
 \code{\link{RemoveExecutable}}, \code{\link{ClearExecutables}}
}

