% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{AddFile}
\alias{AddFile}
\title{Add a file to the DAX}
\usage{
AddFile(adag, file)
}
\arguments{
\item{adag}{ADAG object}

\item{file}{File object}
}
\value{
The ADAG object with the file appended
}
\description{
Add a file to the DAX
}
\seealso{
\code{\link{ADAG}}, \code{\link{File}}, \code{\link{RemoveFile}}, \code{\link{ClearFiles}}
}

