% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R, R/dag.R, R/dax.R, R/executable.R, R/generic_functions.R, R/job.R, R/transformation.R
\name{AddInvoke.ADAG}
\alias{AddInvoke}
\alias{AddInvoke.ADAG}
\alias{AddInvoke.DAG}
\alias{AddInvoke.DAX}
\alias{AddInvoke.Executable}
\alias{AddInvoke.Job}
\alias{AddInvoke.Transformation}
\title{Add an invoke to the object}
\usage{
\method{AddInvoke}{ADAG}(obj, invoke)

\method{AddInvoke}{DAG}(obj, invoke)

\method{AddInvoke}{DAX}(obj, invoke)

\method{AddInvoke}{Executable}(obj, invoke)

AddInvoke(obj, invoke)

\method{AddInvoke}{Job}(obj, invoke)

\method{AddInvoke}{Transformation}(obj, invoke)
}
\arguments{
\item{obj}{Object to append the invoke}

\item{invoke}{The invocation}
}
\value{
The object containing the invoke
}
\description{
Add an invoke to the object
}
\seealso{
\code{\link{ADAG}}

\code{\link{DAG}}

\code{\link{DAX}}

\code{\link{Executable}}

\code{\link{Invoke}}

\code{\link{Job}}

\code{\link{Transformation}}
}

