% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invoke_mixin.R
\name{AddInvokeMixin}
\alias{AddInvokeMixin}
\title{Add invoke to the InvokeMixin object}
\usage{
AddInvokeMixin(invoke.mixin, invoke)
}
\arguments{
\item{invoke.mixin}{InvokeMixin object}

\item{invoke}{invocation to be appended to the list of invocations}
}
\value{
InvokeMixin object with invocation appended to the list of invocations
}
\description{
Add invoke to the InvokeMixin object
}
\seealso{
\code{\link{InvokeMixin}}
}

