% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{AddJob}
\alias{AddJob}
\title{Add a job to the ADAG}
\usage{
AddJob(adag, job)
}
\arguments{
\item{adag}{ADAG object}

\item{job}{Job object}
}
\value{
ADAG object with the job appended
}
\description{
Add a job to the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{RemoveJob}}
}

