% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{AddTransformation}
\alias{AddTransformation}
\title{Add a transformation to the ADAG}
\usage{
AddTransformation(adag, transformation)
}
\arguments{
\item{adag}{ADAG object}

\item{transformation}{Transformation object}
}
\value{
The ADAG object with the transformation appended
}
\description{
Add a transformation to the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{Transformation}},
 \code{\link{RemoveTransformation}}, \code{\link{ClearTransformations}}
}

