% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{AppendToList}
\alias{AppendToList}
\title{Append a value to a list}
\usage{
AppendToList(list, value)
}
\arguments{
\item{list}{List of element to where the value will be appended}

\item{value}{Value to be added to the list}
}
\value{
List with value appended
}
\description{
Append a value to a list
}

