% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{ClearDependencies}
\alias{ClearDependencies}
\title{Remove all dependencies}
\usage{
ClearDependencies(adag)
}
\arguments{
\item{adag}{The ADAG object}
}
\value{
The ADAG object with no dependencies
}
\description{
Remove all dependencies
}
\seealso{
\code{\link{ADAG}}, \code{\link{Dependency}},
 \code{\link{Depends}}, \code{\link{AddDependency}},
 \code{\link{RemoveDependency}}
}

