% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{ClearExecutables}
\alias{ClearExecutables}
\title{Remove all executables}
\usage{
ClearExecutables(adag)
}
\arguments{
\item{adag}{ADAG object}
}
\value{
The ADAG object with no executables
}
\description{
Remove all executables
}
\seealso{
\code{\link{ADAG}}, \code{\link{Executable}},
 \code{\link{AddExecutable}}, \code{\link{RemoveExecutable}}
}

