% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{ClearFiles}
\alias{ClearFiles}
\title{Remove all files}
\usage{
ClearFiles(adag)
}
\arguments{
\item{adag}{ADAG object}
}
\value{
The ADAG object with no files
}
\description{
Remove all files
}
\seealso{
\code{\link{ADAG}}, \code{\link{File}}, \code{\link{AddFile}}, \code{\link{RemoveFile}}
}

