% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invoke_mixin.R
\name{ClearInvokes}
\alias{ClearInvokes}
\title{Remove all Invoke objects}
\usage{
ClearInvokes(invoke.mixin)
}
\arguments{
\item{invoke.mixin}{InvokeMixin object}
}
\value{
InvokeMixin with no invocations
}
\description{
Remove all Invoke objects
}
\seealso{
\code{\link{InvokeMixin}}
}

