% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{ClearJobs}
\alias{ClearJobs}
\title{Remove all jobs}
\usage{
ClearJobs(adag)
}
\arguments{
\item{adag}{ADAG object}
}
\value{
The ADAG object with no jobs
}
\description{
Remove all jobs
}
\seealso{
\code{\link{ADAG}}, \code{\link{RemoveJob}}, \code{\link{AddJob}}
}

