% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{ClearTransformations}
\alias{ClearTransformations}
\title{Remove all transformations}
\usage{
ClearTransformations(adag)
}
\arguments{
\item{adag}{ADAG object}
}
\value{
The ADAG object with no transformations
}
\description{
Remove all transformations
}
\seealso{
\code{\link{ADAG}}, \code{\link{Transformation}},
 \code{\link{AddTransformation}}, \code{\link{RemoveTransformation}}
}

