% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dag.R
\name{DAG}
\alias{DAG}
\title{This job represents a sub-DAG that will be executed by the workflow}
\usage{
DAG(file, id = NULL, node.label = NULL)
}
\arguments{
\item{file}{The logical name of the DAG file, or the DAG File object}

\item{id}{The ID of the DAG job [default: autogenerated]}

\item{node.label}{The label for this job to use in graphing}
}
\value{
The sub-DAG job
}
\description{
This job represents a sub-DAG that will be executed by the workflow
}
\details{
The name argument can be either a string, or a \code{File} object. If
it is a \code{File} object, then this job will inherit its name from the
\code{File} and the \code{File} will be added in a \code{<uses>} with \code{transfer=TRUE},
\code{register=FALSE}, and \code{link=input}.
}
\examples{
dagjob1 <- DAG(file="foo.dag")
dagfile <- File("foo.dag")
dagjob2 <- DAG(dagfile)
}

