% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DAX3.Arch}
\alias{DAX3.Arch}
\title{Architecture types}
\format{\preformatted{List of 8
 $ X86    : chr "x86"
 $ X86_64 : chr "x86_64"
 $ PPC    : chr "ppc"
 $ PPC_64 : chr "ppc_64"
 $ IA64   : chr "ia64"
 $ SPARCV7: chr "sparcv7"
 $ SPARCV9: chr "sparcv9"
 $ AMD64  : chr "amd64"
}}
\usage{
DAX3.Arch
}
\description{
Architecture types
}
\seealso{
\code{\link{Executable}}
}
\keyword{datasets}

