% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DAX3.Link}
\alias{DAX3.Link}
\title{Linkage attributes}
\format{\preformatted{List of 5
 $ NONE      : chr "none"
 $ INPUT     : chr "input"
 $ OUTPUT    : chr "output"
 $ INOUT     : chr "inout"
 $ CHECKPOINT: chr "checkpoint"
}}
\usage{
DAX3.Link
}
\description{
Linkage attributes
}
\seealso{
\code{\link{File}}, \code{\link{Executable}}, \code{\link{Uses}}
}
\keyword{datasets}

