% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DAX3.Namespace}
\alias{DAX3.Namespace}
\title{Namespace values recognized by Pegasus}
\format{\preformatted{List of 8
 $ PEGASUS : chr "pegasus"
 $ CONDOR  : chr "condor"
 $ DAGMAN  : chr "dagman"
 $ ENV     : chr "env"
 $ HINTS   : chr "hints"
 $ GLOBUS  : chr "globus"
 $ SELECTOR: chr "selector"
 $ STAT    : chr "stat"
}}
\usage{
DAX3.Namespace
}
\description{
Namespace values recognized by Pegasus
}
\seealso{
\code{\link{Executable}}, \code{\link{Transformation}}, \code{\link{Job}}
}
\keyword{datasets}

