% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DAX3.OS}
\alias{DAX3.OS}
\title{OS types}
\format{\preformatted{List of 5
 $ LINUX  : chr "linux"
 $ SUNOS  : chr "sunos"
 $ AIX    : chr "aix"
 $ MACOS  : chr "macos"
 $ WINDOWS: chr "windows"
}}
\usage{
DAX3.OS
}
\description{
OS types
}
\seealso{
\code{\link{Executable}}
}
\keyword{datasets}

