% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DAX3.Transfer}
\alias{DAX3.Transfer}
\title{Transfer types for uses}
\format{\preformatted{List of 3
 $ FALSE   : chr "false"
 $ OPTIONAL: chr "optional"
 $ TRUE    : chr "true"
}}
\usage{
DAX3.Transfer
}
\description{
Transfer types for uses
}
\seealso{
\code{\link{Executable}}, \code{\link{File}}
}
\keyword{datasets}

