% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DAX3.When}
\alias{DAX3.When}
\title{Job states for notifications}
\format{\preformatted{List of 6
 $ NEVER     : chr "never"
 $ START     : chr "start"
 $ ON_ERROR  : chr "on_error"
 $ ON_SUCCESS: chr "on_success"
 $ AT_END    : chr "at_end"
 $ ALL       : chr "all"
}}
\usage{
DAX3.When
}
\description{
Job states for notifications
}
\seealso{
\code{\link{Job}}, \code{\link{DAX}}, \code{\link{DAG}}, \code{\link{Invoke}}
}
\keyword{datasets}

