% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{Depends}
\alias{Depends}
\title{Add a dependency to the workflow}
\usage{
Depends(adag, child, parent, edge.label = NULL)
}
\arguments{
\item{adag}{The ADAG object}

\item{child}{The child job/dax/dag or id}

\item{parent}{The parent job/dax/dag or id}

\item{edge.label}{A label for the edge (optional)}
}
\value{
The ADAG object with the dependency appended
}
\description{
Add a dependency to the workflow
}
\seealso{
\code{\link{ADAG}}, \code{\link{Dependency}}, \code{\link{AddDependency}}
}

