% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/element.R
\name{Escape}
\alias{Escape}
\title{Escape special characters in XML}
\usage{
Escape(text)
}
\arguments{
\item{text}{Text to be escaped}
}
\value{
Escaped special character
}
\description{
Escape special characters in XML
}

