% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/executable.R
\name{Executable}
\alias{Executable}
\title{An entry for an executable in the DAX-level replica catalog}
\usage{
Executable(name, namespace = NULL, version = NULL, arch = NULL,
  os = NULL, osrelease = NULL, osversion = NULL, glibc = NULL,
  installed = NULL)
}
\arguments{
\item{name}{Logical name of executable}

\item{namespace}{Executable namespace}

\item{version}{Executable version}

\item{arch}{Architecture that this exe was compiled for}

\item{os}{Name of os that this exe was compiled for}

\item{osrelease}{Release of os that this exe was compiled for}

\item{osversion}{Version of os that this exe was compiled for}

\item{glibc}{Version of glibc this exe was compiled against}

\item{installed}{Is the executable installed (true), or stageable (false)}
}
\value{
The executable object for the program
}
\description{
An entry for an executable in the DAX-level replica catalog
}
\examples{
grep <- Executable("grep")
grep <- Executable(namespace="os",name="grep",version="2.3")
grep <- Executable(namespace="os",name="grep",version="2.3",arch=DAX3.Arch$X86)
grep <- Executable(namespace="os",name="grep",version="2.3",arch=DAX3.Arch$X86,os=DAX3.OS$LINUX)
}
\seealso{
\code{\link{AddExecutable}}
}

