% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{GetJob}
\alias{GetJob}
\title{Get a Job/DAG/DAX}
\usage{
GetJob(adag, jobid)
}
\arguments{
\item{adag}{ADAG object}

\item{jobid}{Job identification}
}
\value{
Job/DAG/DAX object
}
\description{
Get a Job/DAG/DAX
}
\seealso{
\code{\link{ADAG}}, \code{\link{HasJob}}
}

