% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{HasDependency}
\alias{HasDependency}
\title{Check to see if dependency exists}
\usage{
HasDependency(adag, dep)
}
\arguments{
\item{adag}{The ADAG object}

\item{dep}{The dependency object}
}
\value{
If the ADAG contains the dependency
}
\description{
Check to see if dependency exists
}
\seealso{
\code{\link{ADAG}}, \code{\link{Dependency}}
}

