% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{HasExecutable}
\alias{HasExecutable}
\title{Check if executable is in this ADAG}
\usage{
HasExecutable(adag, executable)
}
\arguments{
\item{adag}{ADAG object}

\item{executable}{Executable object}
}
\value{
If the executable is in the ADAG
}
\description{
Check if executable is in this ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{Executable}}
}

