% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{HasFile}
\alias{HasFile}
\title{Check to see if file is in the ADAG}
\usage{
HasFile(adag, file)
}
\arguments{
\item{adag}{ADAG object}

\item{file}{File object}
}
\value{
If the ADAG object contains the file
}
\description{
Check to see if file is in the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{File}}
}

