% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invoke_mixin.R
\name{HasInvoke}
\alias{HasInvoke}
\title{Test whether an invocation is already appended to the InvokeMixin object.}
\usage{
HasInvoke(invoke.mixin, invoke)
}
\arguments{
\item{invoke.mixin}{InvokeMixin object}

\item{invoke}{invocation to be tested}
}
\value{
if the InvokeMixin object has the invocation
}
\description{
Test whether an invocation is already appended to the InvokeMixin object.
}
\seealso{
\code{\link{InvokeMixin}}
}

