% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{HasJob}
\alias{HasJob}
\title{Test to see if job is in this ADAG}
\usage{
HasJob(adag, job)
}
\arguments{
\item{adag}{ADAG object}

\item{job}{Job/DAG/DAX object}
}
\value{
If the Job/DAG/DAX is in the ADAG
}
\description{
The job parameter can be an object or a job ID.
}
\seealso{
\code{\link{ADAG}}, \code{\link{GetJob}}
}

