% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{HasTransformation}
\alias{HasTransformation}
\title{Check to see if transformation is in the ADAG}
\usage{
HasTransformation(adag, transformation)
}
\arguments{
\item{adag}{ADAG object}

\item{transformation}{Transformation object}
}
\value{
If the ADAG has the transformation
}
\description{
Check to see if transformation is in the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{Transformation}}
}

