% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invoke_mixin.R
\name{InvokeMixin}
\alias{InvokeMixin}
\title{Manage invocations}
\usage{
InvokeMixin()
}
\value{
InvokeMixin object with an empty list of invocations
}
\description{
Manage invocations
}
\seealso{
\code{\link{AddInvoke}}, \code{\link{HasInvoke}}, \code{\link{RemoveInvoke}},
 \code{\link{ClearInvokes}}, \code{\link{InvokeExecutable}}
}

