% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{NextJobID}
\alias{NextJobID}
\title{Get an autogenerated ID for the next job}
\usage{
NextJobID(adag)
}
\arguments{
\item{adag}{ADAG object}
}
\value{
DAX object with updated sequence number and the next.id in list format: \code{list(ADAG, next.id)}
}
\description{
Get an autogenerated ID for the next job
}
\seealso{
\code{\link{ADAG}}
}

