% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pfn.R
\name{PFN}
\alias{PFN}
\title{A physical file name. Used to provide URLs for files and executables in the DAX-level replica catalog.}
\usage{
PFN(url, site = "local")
}
\arguments{
\item{url}{The url of the file}

\item{site}{The name of the site}
}
\value{
The PFN object with the URL and site
}
\description{
A physical file name. Used to provide URLs for files and executables in the DAX-level replica catalog.
}
\details{
PFNs can be added to \code{File} and \code{Executable}.
}
\examples{
PFN('http://site.com/path/to/file.txt', 'site')
PFN('http://site.com/path/to/file.txt', site='site')
PFN('http://site.com/path/to/file.txt')
}
\seealso{
\code{\link{AddPFN}}, \code{\link{File}}, \code{\link{Executable}}
}

