% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{RemoveDependency}
\alias{RemoveDependency}
\title{Remove dependency from workflow}
\usage{
RemoveDependency(adag, dep)
}
\arguments{
\item{adag}{The ADAG object}

\item{dep}{The dependency object}
}
\value{
The ADAG object without the dependency
}
\description{
Remove dependency from workflow
}
\seealso{
\code{\link{ADAG}}, \code{\link{Dependency}},
 \code{\link{Depends}}, \code{\link{AddDependency}}
}

