% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{RemoveExecutable}
\alias{RemoveExecutable}
\title{Remove executable from the ADAG}
\usage{
RemoveExecutable(adag, executable)
}
\arguments{
\item{adag}{ADAG object}

\item{executable}{Executable object}
}
\value{
The ADAG object without the executable
}
\description{
Remove executable from the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{Executable}},
 \code{\link{AddExecutable}}, \code{\link{ClearExecutables}}
}

