% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{RemoveFile}
\alias{RemoveFile}
\title{Remove file from this ADAG}
\usage{
RemoveFile(adag, file)
}
\arguments{
\item{adag}{ADAG object}

\item{file}{File object}
}
\value{
The ADAG object without the file
}
\description{
Remove file from this ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{File}}, \code{\link{AddFile}}, \code{\link{ClearFiles}}
}

