% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invoke_mixin.R
\name{RemoveInvoke}
\alias{RemoveInvoke}
\title{Remove an invocation from the InvokeMixin object}
\usage{
RemoveInvoke(invoke.mixin, invoke)
}
\arguments{
\item{invoke.mixin}{InvokeMixin object}

\item{invoke}{invocation to be removed}
}
\value{
InvokeMixin object without the removed invocation
}
\description{
Remove an invocation from the InvokeMixin object
}

