% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{RemoveJob}
\alias{RemoveJob}
\title{Remove job from the ADAG}
\usage{
RemoveJob(adag, job)
}
\arguments{
\item{adag}{ADAG object}

\item{job}{Job/DAG/DAX object}
}
\value{
The ADAG object without the Job/DAG/DAX
}
\description{
Remove job from the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{AddJob}}, \code{\link{ClearJobs}}
}

