% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{RemoveTransformation}
\alias{RemoveTransformation}
\title{Remove transformation from the ADAG}
\usage{
RemoveTransformation(adag, transformation)
}
\arguments{
\item{adag}{ADAG object}

\item{transformation}{Transformation object}
}
\value{
The ADAG object without the transformation
}
\description{
Remove transformation from the ADAG
}
\seealso{
\code{\link{ADAG}}, \code{\link{Transformation}},
 \code{\link{AddTransformation}}, \code{\link{ClearTransformations}}
}

