% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R, R/generic_functions.R, R/transformation.R
\name{ToXML.ADAG}
\alias{ToXML}
\alias{ToXML.ADAG}
\alias{ToXML.Transformation}
\title{Get the XML string for the object}
\usage{
\method{ToXML}{ADAG}(obj)

ToXML(obj)

\method{ToXML}{Transformation}(obj)
}
\arguments{
\item{obj}{Object to parse as XML}
}
\value{
The XML string for the object
}
\description{
Get the XML string for the object


}
\details{
For \code{ADAG}: This is primarily intended for testing. If you have a large
ADAG you should use writeXML instead.
}
\seealso{
\code{\link{ADAG}}, \code{\link{WriteXML}}
}

