% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{WriteXML}
\alias{WriteXML}
\title{Write the ADAG as XML to a stream}
\usage{
WriteXML(adag, out)
}
\arguments{
\item{adag}{The ADAG object}

\item{out}{The stream object (e.g., \code{stdout()}, or a filename)}
}
\description{
Write the ADAG as XML to a stream
}
\examples{
dax <- ADAG('diamond')
WriteXML(dax, stdout())
WriteXML(dax, 'diamond.dax')
}
\seealso{
\code{\link{ADAG}}
}

