% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adag.R
\name{WriteXMLFile}
\alias{WriteXMLFile}
\title{Write the ADAG to an XML file}
\usage{
WriteXMLFile(adag, filename)
}
\arguments{
\item{adag}{The ADAG object}

\item{filename}{Name of the file}
}
\description{
Write the ADAG to an XML file
}
\seealso{
\code{\link{ADAG}}, \code{\link{WriteXML}}
}

